refactor: extract status bar rendering into ui/status_bar.rs (#357)
Second slice of the ui.rs decomposition. Moved draw_status_bar
(~155 lines) into its own submodule. Status bar reads only App
state (mode, connection, conversation, scroll, sync) and renders
into a single Rect with no callbacks - clean self-contained extraction.

ui/mod.rs: 5140 -> 4984 lines.
